fix: move base query DTO to core DTOs folder
