+++ packages/core/src/lib/accounting-template/queries/accounting-template.query.ts
@@ -1,6 +1,6 @@
 import { IQuery } from '@nestjs/cqrs';
 import { AccountingTemplate } from './../accounting-template.entity';
-import { BaseQueryDTO } from '../../core/crud/pagination-params';
+import { BaseQueryDTO } from '../../core/dto/base-query.dto';
 
 export class AccountingTemplateQuery implements IQuery {
 	static readonly type = '[Accounting Template] Query All';
