+++ packages/core/src/lib/api-call-log/dto/api-call-log-filter.dto.ts
@@ -1,7 +1,7 @@
 import { ApiPropertyOptional, IntersectionType, PickType } from '@nestjs/swagger';
 import { IsDateString, IsNumber, IsOptional, IsUUID } from 'class-validator';
 import { ID } from '@gauzy/contracts';
-import { BaseQueryDTO } from '../../core/crud/pagination-params';
+import { BaseQueryDTO } from '../../core/dto/base-query.dto';
 import { TenantOrganizationBaseDTO } from '../../core/dto/tenant-organization-base.dto';
 import { ApiCallLog } from '../api-call-log.entity';
