+++ packages/core/src/lib/core/crud/index.ts
@@ -2,6 +2,6 @@ export * from './crud.factory';
 export * from './crud.controller';
 export * from './crud.service';
 export * from './icrud.controller';
-export * from './pagination-params';
+export * from '../dto/base-query.dto';
 export * from './tenant-aware-crud.service';
 export * from './try-request';
