+++ packages/core/src/lib/core/entities/collection/base.collection.ts
@@ -1,4 +1,4 @@
-import { PlainObject } from './collection.types';
+import { PlainObject } from '@gauzy/contracts';
 
 /**
  *	Base class for collections
