--- packages/core/src/lib/core/entities/collection/collection.types.ts
@@ -1,4 +1,0 @@
-/**
- *	Base class for collections
- */
-export type PlainObject = { [key: string]: any };
