+++ packages/core/src/lib/email-template/queries/email-template.query.ts
@@ -1,6 +1,6 @@
 import { IQuery } from '@nestjs/cqrs';
 import { EmailTemplate } from './../email-template.entity';
-import { BaseQueryDTO } from './../../core/crud/pagination-params';
+import { BaseQueryDTO } from '../../core/dto/base-query.dto';
 
 export class EmailTemplateQuery implements IQuery {
 	static readonly type = '[Email Template] Query All';
